requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2
django-debug-toolbar==4.0.0
django-filter==23.2
django-rest-framework-nested==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Markdown==3.4.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.22.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.4.0
swagger-spec-validator==3.0.3
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.25.11
zipp==3.16.0